+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - support for Delta Chat contact colors
 - style pinned chats in chat list
+- `/pin` and `/unpin` commands to pin/unpin chats
 
 ## [0.5.0]